Backend/sniply_captions_generator/requirements.txt
av==14.4.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
ctranslate2==4.6.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
llvmlite==0.44.0
mpmath==1.3.0
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
packaging==25.0
protobuf==6.31.1
pyee==13.0.0
pysubs2==1.8.0
python-ffmpeg==2.0.12
PyYAML==6.0.2
requests==2.32.4
resampy==0.4.3
sympy==1.14.0
tokenizers==0.21.2
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
